rename benchmark phase to benchmark scenario
In the previous message definition benchmark phase represented a
complete benchmark. However, "phase" means a part or period in a
process. To better reflect the meaning the BenchmarkPhase has been
renamed to BenchmarkScenario. This also enables to introduce a real
phase of a benchmark (e.g. calibration or execution) in a later commit.