require specification of the phase to terminate
The specification of the benchmark phase is especially important in
FBMs. To finish the FBM, the robot should continuously send out the
feedback message until the phase in the benchmark state changes. To
avoid accidentally finishing the next phase via a burst of "old"
feedback messages, the desired phase that should be terminated is
included in the benchmark feedback.